+++ filesystem/android/src/main/java/com/capacitorjs/plugins/filesystem/FilesystemPlugin.java
@@ -3,6 +3,7 @@ package com.capacitorjs.plugins.filesystem;
 import android.Manifest;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
+import android.os.Build;
 import com.capacitorjs.plugins.filesystem.exceptions.CopyFailedException;
 import com.capacitorjs.plugins.filesystem.exceptions.DirectoryExistsException;
 import com.capacitorjs.plugins.filesystem.exceptions.DirectoryNotFoundException;
@@ -18,6 +19,8 @@ import com.getcapacitor.annotation.Permission;
 import com.getcapacitor.annotation.PermissionCallback;
 import java.io.*;
 import java.nio.charset.Charset;
+import java.nio.file.Files;
+import java.nio.file.attribute.BasicFileAttributes;
 import org.json.JSONException;
 
 @CapacitorPlugin(
@@ -312,9 +315,24 @@ public class FilesystemPlugin extends Plugin {
             JSObject data = new JSObject();
             data.put("type", fileObject.isDirectory() ? "directory" : "file");
             data.put("size", fileObject.length());
-            data.put("ctime", null);
             data.put("mtime", fileObject.lastModified());
             data.put("uri", Uri.fromFile(fileObject).toString());
+
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                try {
+                    BasicFileAttributes attr = Files.readAttributes(fileObject.toPath(), BasicFileAttributes.class);
+
+                    // use whichever is the oldest between creationTime and lastAccessTime
+                    if (attr.creationTime().toMillis() < attr.lastAccessTime().toMillis()) {
+                        data.put("ctime", attr.creationTime().toMillis());
+                    } else {
+                        data.put("ctime", attr.lastAccessTime().toMillis());
+                    }
+                } catch (Exception ex) {}
+            } else {
+                data.put("ctime", null);
+            }
+
             call.resolve(data);
         }
     }