+++ push-notifications/src/definitions.ts
@@ -6,10 +6,11 @@ export type PresentationOption = 'badge' | 'sound' | 'alert';
 
 declare module '@capacitor/cli' {
   export interface PluginsConfig {
+    /**
+     * On iOS you can configure the way the push notifications are displayed when the app is in foreground.
+     */
     PushNotifications?: {
       /**
-       * Configure the way push notifications are displayed when the app is in foreground.
-       *
        * This is an array of strings you can combine. Possible values in the array are:
        *   - `badge`: badge count on the app icon is updated (default value)
        *   - `sound`: the device will ring/vibrate when the push notification is received
@@ -20,6 +21,7 @@ declare module '@capacitor/cli' {
        * Only available for iOS.
        *
        * @since 1.0.0
+       * @example ["badge", "sound", "alert"]
        */
       presentationOptions: PresentationOption[];
     };